some clean up fixes in scripts

+++ 2org-2ch/stop.sh
@@ -1,16 +1,16 @@
-#!/bin/bash
+#!/bin/bash -e
 #
 # Copyright IBM Corp All Rights Reserved
 #
 # SPDX-License-Identifier: Apache-2.0
 #
 # Exit on first error, print all commands.
-set -e
+. ./.env
 
 # Shut down the Docker containers that might be currently running.
 
 CONTAINERS=$(docker ps -q)
-if [ -n "$CONTAINERS" ]; then
+if [ -n "$CONTAINERS" -a -e $DOCKER_CONFIG_FILE ]; then
     echo "===> Stopping docker containers."
     docker-compose -f docker-compose.yml stop
     echo "===> Stopped."
